renamed gete_pass to vehicle_pass and migrated all the models
