--- gate_pass/projectsite/gate_pass/admin.py
@@ -1,56 +1,0 @@
-from django.contrib import admin
-from .models import (
-    CustomUser, 
-    UserProfile, SecurityProfile, CashierProfile, AdminProfile,
-    Vehicle, VehiclePass, Registration, 
-    RegistrationStatus, PaymentStatus, 
-    Notification, Announcement
-) 
-
-@admin.register(CustomUser)
-class CustomUserAdmin(admin.ModelAdmin):
-    list_display = ('id', 'corporate_email', 'role')
-
-@admin.register(UserProfile)
-class UserProfileAdmin(admin.ModelAdmin):
-    list_display = ('id', 'user', 'full_name', 'phone_number')
-
-@admin.register(SecurityProfile)
-class SecurityProfileAdmin(admin.ModelAdmin):
-    list_display = ('id', 'user', 'assigned_area')
-
-@admin.register(CashierProfile)
-class CashierProfileAdmin(admin.ModelAdmin):
-    list_display = ('id', 'user', 'cashier_id')
-
-@admin.register(AdminProfile)
-class AdminProfileAdmin(admin.ModelAdmin):
-    list_display = ('id', 'user', 'admin_level')
-
-@admin.register(Vehicle)
-class VehicleAdmin(admin.ModelAdmin):
-    list_display = ('id', 'owner', 'plate_number', 'vehicle_type')
-
-@admin.register(VehiclePass)
-class VehiclePassAdmin(admin.ModelAdmin):
-    list_display = ('id', 'vehicle', 'valid_from', 'valid_until')
-
-@admin.register(Registration)
-class RegistrationAdmin(admin.ModelAdmin):
-    list_display = ('id', 'user', 'vehicle', 'status', 'date_registered')
-
-@admin.register(RegistrationStatus)
-class RegistrationStatusAdmin(admin.ModelAdmin):
-    list_display = ('id', 'status_name')
-
-@admin.register(PaymentStatus)
-class PaymentStatusAdmin(admin.ModelAdmin):
-    list_display = ('id', 'status_name')
-
-@admin.register(Notification)
-class NotificationAdmin(admin.ModelAdmin):
-    list_display = ('id', 'recipient', 'message', 'date_sent')
-
-@admin.register(Announcement)
-class AnnouncementAdmin(admin.ModelAdmin):
-    list_display = ('id', 'title', 'date_posted')
