+++ projectsite/vehicle_pass/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class GatePassConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'gate_pass'
+    name = 'vehicle_pass'
